+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
 
 - Incorrect project digest queries
   [#768](https://github.com/OpenFn/Lightning/issues/768)]
+- Fix issue when purging deleted users
+  [#747](https://github.com/OpenFn/Lightning/issues/747)
 - Generate a random name for Workflows when creating one via the UI.
   [#828](https://github.com/OpenFn/Lightning/issues/828)